Module guard::api::v1::pow::verify_pow

PoW Verification module

Structs

ValidationToken
validation token that clients receive as proof for submiting valid PoW

verify_pow
route handler that verifies PoW and issues a solution token if verification is successful